// ClientApp/src/app/tools-management/models/Procurement.ts
export interface Procurement {
    procurementId: number;
    demandId: number;
    demandTypeId:number;
    itemCategoryId:number;
    itemDetailId: number;
    procurementStatusId: number;
    principalNameId: number;
    manufactureId:number;
    localAgentId: number;
    supplierId: number;
    supplierAId: number;
    supplierMId: number;
    cstTecId: number;
    partOfShipmentId: number;
    sparesCategoryId:number;
    departmentNameId: number;
    departmentName:string;
    itemDetail:string;
    itemName:string;
    cstTec:string;
    supplier:string;
    tenderNumber: string;
    workOrderDate:Date;
    dateOfTenderFloat: Date;
    tenderopeningDate: Date;
    tenderPublishDate: Date;
    tenderNotice: string;
    tenderSpecification: string;
    financialApproval: string;
    workOrder: string;
    dateOfDelivery: Date;
    unitPrice: string;
    qty: string;
    remarks: string;
    procurementDocument: string;
    status: number;
    menuPosition: number;
    isActive: boolean;   
}